bug with def_arr type (Z_Realloc do not fill additional region of memory beyond the original size with zero)

+++ prboom2/src/e6y_launcher.c
@@ -1133,7 +1133,7 @@ BOOL CALLBACK LauncherClientCallback (HWND hDlg, UINT message, WPARAM wParam, LP
         }
       }
       
-     if (i == numwadfiles)
+     if ((size_t)i == numwadfiles)
       {
         if (SendMessage(launcher.listHistory, CB_SETCURSEL, 0, 0) != CB_ERR)
         {
